LayoutTests/svg/animations/svgtransform-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test accumulate=sum animation on SVGAnimateTransform.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 0
PASS rect.transform.animVal.getItem(0).matrix.d is 0
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 1
PASS rect.transform.animVal.getItem(0).matrix.d is 1
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 2
PASS rect.transform.animVal.getItem(0).matrix.d is 2
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 2
PASS rect.transform.animVal.getItem(0).matrix.d is 2
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 3
PASS rect.transform.animVal.getItem(0).matrix.d is 3
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 4
PASS rect.transform.animVal.getItem(0).matrix.d is 4
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 4
PASS rect.transform.animVal.getItem(0).matrix.d is 4
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 5
PASS rect.transform.animVal.getItem(0).matrix.d is 5
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 6
PASS rect.transform.animVal.getItem(0).matrix.d is 6
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 6
PASS rect.transform.animVal.getItem(0).matrix.d is 6
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 6
PASS rect.transform.animVal.getItem(0).matrix.d is 6
PASS rect.transform.baseVal.numberOfItems is 0
PASS rect.transform.animVal.numberOfItems is 1
PASS rect.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS rect.transform.animVal.getItem(0).matrix.a is 6
PASS rect.transform.animVal.getItem(0).matrix.d is 6
PASS rect.transform.baseVal.numberOfItems is 0
PASS successfullyParsed is true

TEST COMPLETE
